Create README.md
There are 2 files for creating the neural network. The 'DeepNN_Hardcoded.py', as the name suggests, creates the NN without 'for' loops. The parameters, the weights and the biases of each layers are hardcoded so it can be more clear how the network propagates and how it calculates its derivatives. The second file 'DeepNN_final.py' uses 'for' loops in order to create the layers.

I have left as comments all the printing commands in the source code just in case. I have tried to comment most steps of the data flow so it can be as informative as possible to someone who wants start studying NNs.